wrote draft
Everything works except for the option to modify the file name to
reflect each branch. To make this work, I might need to use a command
other than shutil.COPY. Maybe there's an option to copy and rename? Or
maybe you can use the copy command if you specify a file as a
destination rather than just a folder?