debugging when modifyOption == 'y'
0) It turns out that you can get shutil.copy () to rename files if you
specify a full file name (rather than just a folder) as the destination.

1) Currently, it adds the optionalNameMod string onto the very end of
the file name (after the file type.) I don't want the file to be
'FractionsProject.sbFichterJ', though. I want it to be
'FractionsProjectFichterJ.sb'. I'll research how to insert characters in
the middle of a string.

2) I should also investigate what's going on with ApplesB. I think it
behaves differently from the other test folders. Is it because it's the
first folder?

3) I commented out many of the questions the program asks the user and
specified some default answers. This should speed up testing.