tweak to prompt text
Changed prompt text to give a more accurate example of how the file
names could be modified with usernames. "FractionsProjectFichterJ.doc,"
for example.  I tried to add a line where you use slashformat() on the
name and path of the text file, but I realized this adds an improper
slash after ".txt." I tried entering some path names with capital
letters and discovered--on a Windows system, at least--that this doesn't
causse any problems.